+++ app/spike/internal/net/acl/list.go
@@ -7,7 +7,9 @@ package acl
 import (
 	"encoding/json"
 	"errors"
+
 	"github.com/spiffe/go-spiffe/v2/workloadapi"
+
 	"github.com/spiffe/spike/app/spike/internal/net/api"
 	"github.com/spiffe/spike/internal/auth"
 	"github.com/spiffe/spike/internal/entity/data"